fix(windows): macOS 호출부가 usize 인데 공유 인자가 u32 였다 — 뿌리를 막는다
§2m.47. CI 가 세 번째로 같은 뿌리에서 빨강이었다. `notification_unread` 는 `usize` 인데 공유 함수가
`u32` 를 받았다. 값이 아니라 개수이므로 공유 쪽을 넓혔다(캐스트를 끼우지 않는다).

세 번이면 뿌리를 막는다. 셋 다 원인이 하나다 — Windows `zig build` 는 `src/platform/macos/**` 를
컴파일조차 하지 않는다. §2m.45 가 찾은 커밋별 게이트는 `zig ast-check` 라 파싱까지만 보고, 이번
같은 타입 오류는 그냥 통과한다.

`tools/check-macos-typecheck.sh` 를 만들었다. CI 의 `file explorer macOS product path` 잡이
컴파일하는 것과 같은 루트(`app_host_abi.zig`)를 `-target aarch64-macos` 로 의미 분석만 돌린다
(`--test-no-exec -fno-emit-bin`). 바이너리를 안 내니 macOS SDK 도 프레임워크도 필요 없다.

실측으로 과거 실패 셋이 전부 로컬에서 재현된다 — 이번 `u32/usize` 는 CI 와 같은 줄 번호까지 같고,
§2m.45 의 별칭 누락도 잡는다. 고친 트리에서는 OK 다.

루트를 test 로 잡는 이유: Zig 는 지연 분석이라 아무도 안 부르는 코드는 안 본다.

게이트에 안 걸었다 — 그것은 결정이라 사용자 몫이다. 지금은 macOS 파일을 건드린 슬라이스가 푸시
전에 손으로 돌리는 도구다. 링크는 안 보므로 진짜 macOS 러너를 대체하지 않는다.

+++ src/platform/cell_text.zig
@@ -1000,7 +1000,7 @@ pub fn appendSidebarBellAndBadge(
     cells: *std.ArrayList(renderer.DrawCell),
     bell_col: u16,
     fg: terminal.Color,
-    unread: u32,
+    unread: usize,
     round_badge: bool,
 ) !void {
     try cells.append(allocator, .{ .row = 0, .col = bell_col, .codepoint = icons.codepoint(.bell), .width = 2, .style = .{ .foreground = fg } });
@@ -1034,7 +1034,7 @@ pub fn appendSidebarHeaderIcons(
     cells: *std.ArrayList(renderer.DrawCell),
     cols: u16,
     fg: terminal.Color,
-    unread: u32,
+    unread: usize,
 ) !bool {
     if (cols < sidebar_header_min_cols) return false;
     const sb = chrome.components.sidebar;
